Cleanup chunk servers when dropping dependencies
Chunk server mappings are now cleaned up when dropping chunk tables
and foreign servers. In particular, this works even when such objects
are removed as a result of cascading deletions of other objects.

Some refactoring has been done to the event trigger handling code in
addition to adding support for new event objects.