Support pushing partial aggregates to data nodes
This change modifies the timescale_fdw to allow aggregates on a subset
of partitioning dimensions to be pushed down to data node.  The
deparsing code has also been modified to wrap the pushed aggregate in
a _timescaledb_internal.partialize_agg call, which will return the
proper array of values which postgres can combine into the finalized
value.