Fix crash when creating distributed hypertable
If a distributed hypertable is created but one or more of the servers
are not defined, a crash will result. The reason for this is that an
attempt is made to print an error message using fromctx.cstate in
ChunkCopyState. Since the error message is printed out in the last
phase of the command, the fromctx union is not set to a cstate and
instead to a scandesc, which causes a segmentation fault.

This commit fixes the issue by using different ErrorContextCallback for
the different phases and pass that to timescaledb_CopyFrom. For the
initial CopyFrom phases, the default CopyFromErrorCallback is used,
while a newly defined error callback is used for the last phase.