Refactor estimation code for remote queries
This change refactors the estimation code for calculating the cost of
scanning remote relations. Estimation is now broken up based on the
type of relation one is estimating the cost of (e.g., base or upper
relations). It also adds better handling of upper relations that
include pathkeys (i.e., it is possible to generate sorted paths for
them).