Improve pushdown handling of time functions
This change allows certain mutable functions to be whitelisted so that
they can be safely pushed down to data nodes.  Additionally, this change
will no longer prevent queries containing the `now` function from being
pushed down to data nodes, but will instead replace the function call
with the transaction start time (which is the same value which would be
used had the query been run solely on the access node).