Change location of user certificates and keys
User certificates and keys for logging into data nodes are stored at
the top level of the `ssl_dir` or in the data directory. This can cause
some confusion since a lot of files with user names resembling existing
configuration files will be created as users are added, so this commit
change the location of the user certificates and keys to be in the
`timescaledb/certs` subdirectory of either the `ssl_dir` or data
directory.

In addition, since user names can contain strange characters (quoted
names are allowed as role names, which can contain anything) the commit
changes the names for certificates and keys to use the MD5 sum as hex
string as base name for the files. This will prevent strange user names
from accessing files outside the certificate directory.

The subdirectory is currently hardcoded.