Allow ALTER TABLE SET on distributed hypertable
This change allows to use ALTER TABLE SET/RESET, SET OIDS and SET
WITHOUT OIDS clauses with a distributed hypertable.

This PR has two effects:

1. It prevents having to copy storage options for foreign table chunks
when their objects are created on the AN. The command updates only root
table options on the AN and passes it for execution on the data nodes.

2 It prevents distributed hypertable chunks to be updated in the
'ddl_command_end' event trigger on AN, because PostgreSQL does not
support altering storage options for foreign tables.